Refactor crashdumping
Rename "exception handling" to "crash dumping" because "exception
handling" sounds like C++ exceptions to me.

Don't rename the config option to avoid a breaking change.

Use the conditional compilation only in a single compilation unit so
that testing the feature flag can be independent of everything else.

Only expose a single function to enable crash dumping, the previous
two-step process didn't offer benefits.
